requirements.txt
# Core
requests>=2.31
pyyaml>=6.0
beautifulsoup4>=4.12
lxml>=5.1
feedparser>=6.0

# CV Reading
PyPDF2>=3.0
python-docx>=1.1

# Matching & NLP
scikit-learn>=1.4
rapidfuzz>=3.6

# Email
jinja2>=3.1

# Scheduling
schedule>=1.2

# LLM Matching
anthropic>=0.40

# Web Dashboard
flask>=3.0
waitress>=3.0
